+++ examples/go_test.go
@@ -25,7 +25,6 @@ func TestGoExample(t *testing.T) {
 		Secrets: map[string]string{
 			"dockerHubPassword": os.Getenv("DOCKER_HUB_PASSWORD"),
 		},
-		ExtraRuntimeValidation: assertSecretsInputIsSecret,
 	}
 
 	integration.ProgramTest(t, &test)
