+++ examples/nodejs_test.go
@@ -37,7 +37,6 @@ func TestNodeExample(t *testing.T) {
 		Secrets: map[string]string{
 			"dockerHubPassword": os.Getenv("DOCKER_HUB_PASSWORD"),
 		},
-		ExtraRuntimeValidation: assertSecretsInputIsSecret,
 	}
 
 	integration.ProgramTest(t, &test)
