+++ examples/python_test.go
@@ -26,7 +26,6 @@ func TestPythonExample(t *testing.T) {
 		Secrets: map[string]string{
 			"dockerHubPassword": os.Getenv("DOCKER_HUB_PASSWORD"),
 		},
-		ExtraRuntimeValidation: assertSecretsInputIsSecret,
 	}
 
 	integration.ProgramTest(t, &test)
